chore(core): improve type hints and simplify code structure

+++ src/domain/delivery/ports/cost_calculator_interface.py
@@ -1,4 +1,5 @@
 import abc
+from decimal import Decimal
 
 from domain.maps.model.value_objects import Address
 from domain.maps.ports.maps_adapter_interface import MapsAdapterInterface
@@ -11,16 +12,16 @@ class DeliveryCostCalculatorAdapterInterface(abc.ABC):
         self.maps_service = maps_service
 
     @abc.abstractmethod
-    async def calculate_cost(self, total_product_cost: float, destination: Address) -> float:
+    async def calculate_cost(self, total_product_cost: Decimal, destination: Address) -> Decimal:
         """Calculate the delivery cost based on product cost and destination."""
         raise NotImplementedError
 
     @abc.abstractmethod
-    async def _large_delivery_calculate_cost(self, destination: Address) -> float:
+    async def _large_delivery_calculate_cost(self, destination: Address) -> Decimal:
         """Specialized calculation for large deliveries."""
         raise NotImplementedError()
 
     @abc.abstractmethod
-    async def _small_delivery_calculate_cost(self, destination: Address) -> float:
+    async def _small_delivery_calculate_cost(self, destination: Address) -> Decimal:
         """Specialized calculation for small deliveries."""
         raise NotImplementedError()
